pyproject.toml
[project]
name = "vibe-cids"
version = "0.1.0"
description = "CIDSeeks collaborative intrusion detection simulation (Evaluation-2 SimPy)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "simpy>=4.0.0",
  "PyYAML>=6.0",
  "numpy>=1.20.0",
  "pandas>=1.3.0",
  "scikit-learn>=1.0.0",
  "scipy>=1.7.0",
  "psutil>=5.8.0",
  "customtkinter>=5.0.0",
  "Pillow>=8.0.0",
  "pywebview>=4.0",
  "tkinterweb>=3.6.0",
  "cryptography>=3.4.0",
  "networkx>=2.8.0",
  "pyvis>=0.3.2",
  "matplotlib>=3.4.0",
  "seaborn>=0.11.0",
  "plotly>=5.0.0",
  "openpyxl>=3.0.0",
  "xlsxwriter>=3.0.0",
  "tkhtmlview",
  "python-dotenv",
]

[project.optional-dependencies]
dev = [
  "pytest>=6.0",
  "pytest-cov>=2.0",
  "flake8>=3.9",
  "mypy>=0.910",
  "black>=21.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
